(Claude) Phase 1: Complete core module migration
Migrated all top-level Python modules into proper onion architecture layers.

Changes:
- Deleted config.py, logger.py, messaging.py, user_message.py from root
- Migrated config to bt_servant_engine/core/config.py with backward-compatible alias
- Migrated logging to bt_servant_engine/core/logging.py
- Moved UserMessage model to bt_servant_engine/core/models.py
- Merged messaging functionality into bt_servant_engine/adapters/messaging.py
- Updated all imports across codebase (52 files)
- Fixed pre-existing syntax error in utils/perf.py (multi-line f-strings)
- Updated db/ module imports for compatibility (Phase 2 will migrate db/ fully)
- Added test environment defaults (LOG_PSEUDONYM_SECRET, ENABLE_ADMIN_AUTH)

Result: Only bt_servant.py and brain.py remain at root level.
All tests pass. Ruff linting passes.
