+++ bt_servant_engine/services/__init__.py
@@ -3,10 +3,9 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional
 
 from bt_servant_engine.core.ports import ChromaPort, MessagingPort, UserStatePort
-
 
 if TYPE_CHECKING:  # pragma: no cover - type narrowing only
     from .intent_router import IntentRouter
