--- logger.py
@@ -1,19 +1,0 @@
-"""Legacy shim exposing logging helpers from ``bt_servant_engine.core``."""
-
-from bt_servant_engine.core.logging import (  # noqa: F401
-    CorrelationIdFilter,
-    bind_correlation_id,
-    correlation_id_context,
-    get_correlation_id,
-    get_logger,
-    reset_correlation_id,
-)
-
-__all__ = [
-    "CorrelationIdFilter",
-    "bind_correlation_id",
-    "correlation_id_context",
-    "get_correlation_id",
-    "get_logger",
-    "reset_correlation_id",
-]
