+++ tests/services/test_intent_router.py
@@ -7,7 +7,6 @@ import asyncio
 import pytest
 
 from brain import IntentType
-
 from bt_servant_engine.services import ServiceContainer, build_default_services
 from bt_servant_engine.services.intent_router import (
     IntentHandlerNotFoundError,
