+++ utils/__init__.py
@@ -1,5 +1,5 @@
 """Utility modules for text, BSB parsing, and keyword helpers."""
 
-from .text_utils import combine_chunks, chop_text
+from .text_utils import chop_text, combine_chunks
 
 __all__ = ["combine_chunks", "chop_text"]
