+++ utils/identifiers.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 
 import base64
-import hmac
 import hashlib
+import hmac
 import os
 from functools import lru_cache
 from typing import Optional
