+++ utils/usfm_to_json.py
@@ -11,15 +11,13 @@ Provides:
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 import json
+import re
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, Iterable, List, Tuple
 
-import re
-
 from .bsb import BOOK_MAP
-
 
 USFM_CODE_TO_BOOK: Dict[str, str] = {
     # OT
